PR feedback updates
