+++ src/v3/test/integration/admin-consent.test.tsx
@@ -16,30 +16,12 @@ import adminConsentResponse from '../../../../playground/mocks/data/idp/idx/cons
 
 describe('admin-consent', () => {
   it('should render form with logo', async () => {
-    const { container, findByRole } = await setup({ mockResponse: adminConsentResponse });
+    const { container, findByRole, queryByAltText } = await setup({ mockResponse: adminConsentResponse });
     const appNameHeading = await findByRole('heading', { level: 2 });
     const groupHeading = await findByRole('heading', { level: 3 });
+    const logo = queryByAltText('Logo for the app');
 
-    expect(appNameHeading.textContent).toBe('Native client');
-    expect(groupHeading.textContent).toBe('Resource and policies');
-
-    expect(container).toMatchSnapshot();
-  });
-
-  it('should render form without logo', async () => {
-    const adminConsentResponseWithoutLogo = {
-      ...adminConsentResponse,
-      app: {
-        ...adminConsentResponse.app,
-        value: { ...adminConsentResponse.app.value, logo: undefined },
-      },
-    };
-    const { container, findByRole } = await setup({
-      mockResponse: adminConsentResponseWithoutLogo,
-    });
-    const appNameHeading = await findByRole('heading', { level: 2 });
-    const groupHeading = await findByRole('heading', { level: 3 });
-
+    expect(logo).toBeDefined();
     expect(appNameHeading.textContent).toBe('Native client');
     expect(groupHeading.textContent).toBe('Resource and policies');
 
@@ -61,8 +43,8 @@ describe('admin-consent', () => {
   it('should send correct payload when cancel is clicked', async () => {
     const { authClient, user, findByText } = await setup({ mockResponse: adminConsentResponse });
 
-    const allowConsentBtn = await findByText('Cancel');
-    await user.click(allowConsentBtn);
+    const cancelBtn = await findByText('Cancel');
+    await user.click(cancelBtn);
     expect(authClient.options.httpRequestClient).toHaveBeenCalledWith(
       ...createAuthJsPayloadArgs('POST', 'idp/idx/consent', {
         consent: false,
